fix issues with vite.config.js not being able to read env variables
