+++ .env.chrome
@@ -1,0 +1,1 @@
+VITE_TARGET_BROWSER=chrome
