+++ .env.firefox
@@ -1,0 +1,1 @@
+VITE_TARGET_BROWSER=firefox
