--- vite.config.cjs
@@ -1,29 +1,0 @@
-import { defineConfig } from "vite";
-import webExtension, { readJsonFile } from "vite-plugin-web-extension";
-
-function generateManifest() {
-  const manifest = readJsonFile("src/manifest.json");
-  const pkg = readJsonFile("package.json");
-  return {
-    version: pkg.version,
-    ...manifest
-  };
-}
-
-export default defineConfig({
-  plugins: [
-    webExtension({
-      manifest: generateManifest,
-      watchFilePaths: ["package.json", "src/manifest.json"],
-      browser: process.env.TARGET_BROWSER || "chrome",
-      webExtConfig: {
-        startUrl: [
-          "https://www.youtube.com/playlist?list=PLAhTBeRe8IhMmRve_rSfAgL_dtEXkKh8Z"
-        ]
-      }
-    })
-  ],
-  build: {
-    sourcemap: "inline"
-  }
-});
